feat (pytorch): autolog plain tensorboard if lightning isn't active (#7627)
* feat (pytorch): autolog plain tensorboard if lightning isn't active

Currently, the "pytorch" flavor isn't very helpful when lightning isn't
installed/used. This PR ensures that scalars + hparams logged to
tensorboard are also logged to mlflow even when no lightning is used.


* fixes for log_every_n_step


* support pytorch autolog even w/o lightning installed


* test for timestamps


* test pytorch autolog *without* lightning installed


* relax requirements for torch autologging


* Update dev/run-python-tests.sh


* dev install for torch


* Update mlflow/ml-package-versions.yml


* fix test for unsupported pytorch versions


* use exising versions for torch in ml-package-versions.yml


* lightning autolog: warn user for unsupported dependency versions


* fix test import for renamed file


* update pytorch autolog docstring


* add dev torch to lightning's dev deps


* clean up imports


* add torch to mock _ML_PACKAGE_VERSIONS


* encode user code dependency on tensorboard required for testing


* reduce size of test matrix


* fix test


* more test fixing


* add tensorboard to test requirements


* remove redundant tensorboard_autolog test call


* handle importerrror in _lightning_autolog outside of the file


* add test for disabled tensorboard patching when in lightning's fit()


* Update mlflow/ml-package-versions.yml


* run_id usage changes suggested in code review


* remove redundant lightning dependencies version checks


* dict() -> {...}


* idem


* remove _get_version


* linting


* docs: lightning version constraint formatting


* _get_run() -> mlflow.active_run()


* enable / disable plain pytorch logging from lightning's fit


* manage run in safe_patch


* lint


* cleaner disabling mechanism for plain pytorch logging


* missing return


* suggested original/queue flush order change


---------
