+++ mlflow/ml-package-versions.yml
@@ -42,6 +42,14 @@ pytorch:
     run: |
       pytest tests/pytorch/test_pytorch_model_export.py tests/pytorch/test_pytorch_metric_value_conversion_utils.py
 
+  autologging:
+    minimum: "1.6.0"
+    maximum: "1.13.1"
+    requirements:
+      ">= 0.0.0": ["tensorboard"]
+    run: |
+      pytest tests/pytorch/test_tensorboard_autolog.py
+
 pytorch-lightning:
   package_info:
     pip_release: "pytorch-lightning"
