+++ mlflow/ml_package_versions.py
@@ -11,6 +11,15 @@ _ML_PACKAGE_VERSIONS = {
             "maximum": "1.2.0"
         }
     },
+    "pytorch": {
+        "package_info": {
+            "pip_release": "torch"
+        },
+        "autologging": {
+            "minimum": "1.6.0",
+            "maximum": "1.13.1"
+        }
+    },
     "pytorch-lightning": {
         "package_info": {
             "pip_release": "pytorch-lightning"
