+++ mlflow/utils/autologging_utils/versioning.py
@@ -16,7 +16,7 @@ FLAVOR_TO_MODULE_NAME_AND_VERSION_INFO_KEY = {
     "tensorflow": ("tensorflow", "tensorflow"),
     "xgboost": ("xgboost", "xgboost"),
     "sklearn": ("sklearn", "sklearn"),
-    "pytorch": ("pytorch_lightning", "pytorch-lightning"),
+    "pytorch": ("torch", "pytorch"),
     "pyspark.ml": ("pyspark", "spark"),
 }
