+++ tests/pytorch/test_pytorch_autolog.py
@@ -16,7 +16,7 @@ from mlflow import MlflowClient
 import mlflow.pytorch
 from mlflow.exceptions import MlflowException
 from mlflow.utils.file_utils import TempDir
-from mlflow.pytorch._pytorch_autolog import _get_optimizer_name
+from mlflow.pytorch._lightning_autolog import _get_optimizer_name
 
 NUM_EPOCHS = 20
 
@@ -325,6 +325,8 @@ def test_pytorch_test_metrics_logged(pytorch_model_tests):
     data = run.data
     assert "test_loss" in data.metrics
     assert "test_acc" in data.metrics
+    # this is logged through SummaryWriter
+    assert "plain_loss" not in data.metrics
 
 
 def test_get_optimizer_name():
